Added web application setup
