+++ core/build.gradle
@@ -1,3 +1,10 @@
+sourceCompatibility = 1.11
+[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
+
+sourceSets.main.java.srcDirs = [ "src/" ]
+sourceSets.test.java.srcDirs = ["test/java"]
+eclipse.project.name = appName + "-core"
+
 jacocoTestReport {
     reports {
         xml.required = false
@@ -5,11 +12,3 @@ jacocoTestReport {
         html.outputLocation = layout.buildDirectory.dir('jacocoHtml')
     }
 }
-
-sourceCompatibility = 1.11
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-
-sourceSets.main.java.srcDirs = [ "src/" ]
-sourceSets.test.java.srcDirs = ["test/java"]
-
-eclipse.project.name = appName + "-core"
