+++ core/src/FreedomGame.gwt.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "http://www.gwtproject.org/doctype/2.8.0/gwt-module.dtd">
+<module>
+	<source path="it/units/sdm/project">
+		<exclude name="utils/MainLogger.java"/>
+		<exclude name="core/FreedomGame.java" />
+		<exclude name="TextInput.java" />
+	</source>
+
+</module>
