+++ core/src/it/units/sdm/project/Position.java
@@ -45,7 +45,7 @@ public class Position implements Comparable<Position> {
     @Override
     public String toString() {
         char column = (char) ('A' + this.column - 1);
-        return String.format("%c%d", column, row);
+        return column + " " + row;
     }
 
     /**
