+++ desktop/build.gradle
@@ -1,6 +1,8 @@
 sourceCompatibility = 1.11
 sourceSets.main.java.srcDirs = [ "src/" ]
 sourceSets.main.resources.srcDirs = ["../assets"]
+sourceSets.test.java.srcDirs = ["test/java"]
+
 
 project.ext.mainClassName = "it.units.sdm.project.DesktopLauncher"
 project.ext.assetsDir = new File("../assets")
