+++ desktop/src/it/units/sdm/project/DesktopLauncher.java
@@ -2,7 +2,6 @@ package it.units.sdm.project;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import it.units.sdm.project.Freedom;
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
 public class DesktopLauncher {
@@ -10,6 +9,6 @@ public class DesktopLauncher {
 		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
 		config.setForegroundFPS(60);
 		config.setTitle("SDM");
-		new Lwjgl3Application(new Freedom(), config);
+		new Lwjgl3Application(new FreedomGame(), config);
 	}
 }
