+++ html/src/it/units/sdm/project/client/HtmlLauncher.java
@@ -1,0 +1,22 @@
+package it.units.sdm.project.client;
+
+import com.badlogic.gdx.ApplicationListener;
+import com.badlogic.gdx.backends.gwt.GwtApplication;
+import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
+import it.units.sdm.project.FreedomGame;
+
+public class HtmlLauncher extends GwtApplication {
+
+        @Override
+        public GwtApplicationConfiguration getConfig () {
+                // Resizable application, uses available space in browser
+                return new GwtApplicationConfiguration(true);
+                // Fixed size application:
+                //return new GwtApplicationConfiguration(480, 320);
+        }
+
+        @Override
+        public ApplicationListener createApplicationListener () {
+                return new FreedomGame();
+        }
+}
