+++ html/webapp/WEB-INF/web.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" ?>
+<web-app>
+</web-app>
