+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'SDM_project'
-include 'desktop', 'core'
+include 'desktop', 'html', 'core'
